Передача координат точек через основной модуль 3wifi.php
