+++ left.html
@@ -5,7 +5,7 @@
 <h4>3WiFi</h4>
 <ul>
 	<li><a href="main.html" target=mainFrame>Главная</a></li>
-	<li><a href="map3.php" target=mainFrame>Карта</a></li>
+	<li><a href="map.php" target=mainFrame>Карта</a></li>
 	<li><a href="find.php" target=mainFrame>Поиск</a></li>
 	<li><a href="upload.php" target=mainFrame>Добавить в базу</a></li>
 	<li><a href="stat.html" target=mainFrame>Статистика</a></li>
