+++ main.html
@@ -11,7 +11,7 @@
 <br><br>
 <table>
 <tr>
-	<td><a href="map3.php" title="Карта"><img src="img/map.png"></a></td>
+	<td><a href="map.php" title="Карта"><img src="img/map.png"></a></td>
 	<td><a href="find.php" title="Поиск"><img src="img/find.png"></a></td>
 </tr>
 <tr>
